Move stateSevice to state utils for better architecture
