+++ src/commands/account/clear.ts
@@ -1,4 +1,4 @@
-import { recordCommand } from '../../state/stateService';
+import stateUtils from '../../utils/state';
 
 import accountUtils from '../../utils/account';
 import type { Command } from '../../../types';
@@ -14,7 +14,7 @@ export default (program: any) => {
         thisCommand.parent.action().name(),
         ...thisCommand.parent.args,
       ];
-      recordCommand(command);
+      stateUtils.recordCommand(command);
     })
     .description('Clear all accounts from the address book')
     .action(() => {
