+++ src/commands/account/delete.ts
@@ -1,4 +1,4 @@
-import { recordCommand } from '../../state/stateService';
+import stateUtils from '../../utils/state';
 import { Logger } from '../../utils/logger';
 
 import accountUtils from '../../utils/account';
@@ -16,7 +16,7 @@ export default (program: any) => {
         thisCommand.parent.action().name(),
         ...thisCommand.parent.args,
       ];
-      recordCommand(command);
+      stateUtils.recordCommand(command);
     })
     .description('Delete an account from the address book')
     .option('-a, --alias <alias>', 'account must have an alias')
