+++ src/commands/account/list.ts
@@ -1,4 +1,4 @@
-import { recordCommand } from '../../state/stateService';
+import stateUtils from '../../utils/state';
 import accountUtils from '../../utils/account';
 import { Logger } from '../../utils/logger';
 
@@ -14,7 +14,7 @@ export default (program: any) => {
         thisCommand.parent.action().name(),
         ...thisCommand.parent.args,
       ];
-      recordCommand(command);
+      stateUtils.recordCommand(command);
     })
     .description('List all accounts in the address book')
     .option('-p, --private', 'Show private keys')
