+++ src/commands/state/download.ts
@@ -1,6 +1,5 @@
-import { recordCommand } from '../../state/stateService';
-import { Logger } from '../../utils/logger';
 import stateUtils from '../../utils/state';
+import { Logger } from '../../utils/logger';
 
 import type { Command } from '../../../types';
 
@@ -14,7 +13,7 @@ export default (program: any) => {
         thisCommand.parent.action().name(),
         ...thisCommand.parent.args,
       ];
-      recordCommand(command);
+      stateUtils.recordCommand(command);
     })
     .description(
       'Download state from a URL and merge it with the current state',
