+++ src/commands/token/associate.ts
@@ -1,5 +1,5 @@
 import { associateToken } from '../../utils/token';
-import { recordCommand } from '../../state/stateService';
+import stateUtils from '../../utils/state';
 import { Logger } from '../../utils/logger';
 
 import type { Command } from '../../../types';
@@ -15,7 +15,7 @@ export default (program: any) => {
         thisCommand.parent.action().name(),
         ...thisCommand.parent.args,
       ];
-      recordCommand(command);
+      stateUtils.recordCommand(command);
     })
     .description('Associate a token with an account')
     .requiredOption(
