+++ src/commands/token/create.ts
@@ -2,11 +2,7 @@ import { TokenCreateTransaction, TokenType, PrivateKey } from '@hashgraph/sdk';
 
 import { myParseInt } from '../../utils/verification';
 import { getSupplyType } from '../../utils/token';
-import {
-  recordCommand,
-  getHederaClient,
-  getNetwork,
-} from '../../state/stateService';
+import stateUtils from '../../utils/state';
 import { Logger } from '../../utils/logger';
 import stateController from '../../state/stateController';
 
@@ -23,7 +19,7 @@ export default (program: any) => {
         thisCommand.parent.action().name(),
         ...thisCommand.parent.args,
       ];
-      recordCommand(command);
+      stateUtils.recordCommand(command);
     })
     .description('Create a new fungible token')
     .requiredOption(
@@ -99,7 +95,7 @@ async function createFungibleToken(
   supplyType: string,
   adminKey: string,
 ): Promise<string> {
-  const client = getHederaClient();
+  const client = stateUtils.getHederaClient();
 
   let tokenId;
   try {
@@ -149,7 +145,7 @@ async function createFungibleToken(
       decimals,
       initialSupply,
       adminKey,
-      network: getNetwork(),
+      network: stateUtils.getNetwork(),
     },
   };
