+++ src/commands/wait.ts
@@ -1,4 +1,4 @@
-import { recordCommand } from '../state/stateService';
+import stateUtils from '../utils/state';
 import { Logger } from '../utils/logger';
 
 import type { Command } from '../../types';
@@ -13,7 +13,7 @@ export default (program: any) => {
         thisCommand.parent.action().name(),
         ...thisCommand.parent.args,
       ];
-      recordCommand(command);
+      stateUtils.recordCommand(command);
     })
     .description('Wait for a specified number of seconds')
     .action((seconds: string) => {
